Found the error: the files being writting are missing almost a third of their supposed data

+++ GluonicLQCD/parallelization/index.cpp
@@ -34,8 +34,8 @@ void Parallel::Index::setNTot(unsigned int NSpatial, unsigned int NTemporal)
      */
     m_NTot.resize(4);
     for (int i = 0; i < 3; i++) {
-        m_NTot[i] = (unsigned int) NSpatial;
+        m_NTot[i] = (unsigned long int) NSpatial;
     }
-    m_NTot[3] = (unsigned int) NTemporal;
+    m_NTot[3] = (unsigned long int) NTemporal;
 }
