+++ GluonicLQCD/parallelization/index.h
@@ -14,15 +14,15 @@ namespace Parallel {
         ~Index();
 
         // Index getter
-        static inline unsigned long int cubeIndex(unsigned int i, unsigned int j, unsigned int k, unsigned int Ni, unsigned int Nj) {
+        static inline unsigned long int cubeIndex(unsigned long int i, unsigned long int j, unsigned long int k, unsigned long int Ni, unsigned long int Nj) {
             return i + Ni*(j + Nj*k);
         }
 
-        static inline unsigned long int getIndex(unsigned int i, unsigned int j, unsigned int k, unsigned int l) {
+        static inline unsigned long int getIndex(unsigned long int i, unsigned long int j, unsigned long int k, unsigned long int l) {
             return i + m_N[0]*(j + m_N[1]*(k + m_N[2]*l));
         }
 
-        static inline unsigned long int getGlobalIndex(unsigned int i, unsigned int j, unsigned int k, unsigned int l) {
+        static inline unsigned long int getGlobalIndex(unsigned long int i, unsigned long int j, unsigned long int k, unsigned long int l) {
             return i + m_NTot[0]*(j + m_NTot[1]*(k + m_NTot[2]*l)); // column-major
         }
